+++ tests/courseformatrenderer_test.php
@@ -93,8 +93,13 @@ final class courseformatrenderer_test extends \advanced_testcase {
     /**
      * Init.
      */
-    protected function init($numsections = 1, $layoutcolumnorientation = 2, $flexiblemodules = 1,
-        $toggleallenabled = 2, $viewsinglesectionenabled = 2): void {
+    protected function init(
+        $numsections = 1,
+        $layoutcolumnorientation = 2,
+        $flexiblemodules = 1,
+        $toggleallenabled = 2,
+        $viewsinglesectionenabled = 2
+    ): void {
         $this->resetAfterTest(true);
 
         set_config('theme', 'boost');
@@ -262,7 +267,7 @@ final class courseformatrenderer_test extends \advanced_testcase {
 
     public function test_topcoll_section(): void {
         $this->init();
-        set_user_preference(togglelib::TOPCOLL_TOGGLE.'_' . $this->course->id, 'Z');
+        set_user_preference(togglelib::TOPCOLL_TOGGLE . '_' . $this->course->id, 'Z');
         set_config('defaultuserpreference', 0, 'format_topcoll');
         set_config('defaulttogglepersistence', 1, 'format_topcoll');
         self::set_property($this->outputus, 'formatresponsive', false);
@@ -287,6 +292,7 @@ final class courseformatrenderer_test extends \advanced_testcase {
                 $onsectionpage, ]),
             'heading' => '<h3 data-for="section_title" data-id="' . $section1->id . '" data-number="1" id="sectionid-' . $section1->id .
                 '-title" class="sectionname">Section 1<div class="cttoggle"> - Toggle</div></h3>',
+            'id' => $section1->id,
             'nomtore' => true,
             'rightcontent' => self::call_method($this->outputus, 'section_right_content', [$section1, $this->course,
                 $onsectionpage, ]),
@@ -376,8 +382,11 @@ final class courseformatrenderer_test extends \advanced_testcase {
             'columnclass' => 'col-sm-12',
             'cscml' => self::call_method($this->outputus, 'course_section_cmlist', [$section]),
             'heading' => '<h3 data-for="section_title" data-id="' . $section->id . '" data-number="1" id="sectionid-' .
-                $section->id . '-title" class="section-title">' . get_string('orphanedactivitiesinsectionno', '',
-                $section->section) . '</h3>',
+                $section->id . '-title" class="section-title">' . get_string(
+                    'orphanedactivitiesinsectionno',
+                    '',
+                    $section->section
+                ) . '</h3>',
             'rightcontent' => self::call_method($this->outputus, 'section_right_content', [$section, $this->course, false]),
             'rtl' => false,
             'sectionid' => $section->id,
@@ -447,7 +456,7 @@ final class courseformatrenderer_test extends \advanced_testcase {
         global $CFG;
 
         $this->init();
-        set_user_preference(togglelib::TOPCOLL_TOGGLE.'_' . $this->course->id, null);
+        set_user_preference(togglelib::TOPCOLL_TOGGLE . '_' . $this->course->id, null);
         set_config('defaultuserpreference', 0, 'format_topcoll');
         set_config('defaulttogglepersistence', 1, 'format_topcoll');
         $section0 = $this->courseformat->get_section(0);
@@ -462,7 +471,7 @@ final class courseformatrenderer_test extends \advanced_testcase {
         $theoutput .= '</ul><ul class="ctopics ctoggled topics row">';
         $theoutput .= self::call_method($this->outputus, 'topcoll_section', [$section1, $this->course, false, null, $toggle]);
         $theoutput .= '</ul>';
-        $theoutput .= '<span id="tcdata" class="d-none" data-onetopic="false" data-onetopictoggle="false"'.
+        $theoutput .= '<span id="tcdata" class="d-none" data-onetopic="false" data-onetopictoggle="false"' .
             ' data-defaulttogglepersistence="true"></span>';
 
         $this->assertEquals($theoutput, $thevalue);
@@ -472,7 +481,7 @@ final class courseformatrenderer_test extends \advanced_testcase {
         global $CFG;
 
         $this->init(1, 1);
-        set_user_preference(togglelib::TOPCOLL_TOGGLE.'_' . $this->course->id, 'Z');
+        set_user_preference(togglelib::TOPCOLL_TOGGLE . '_' . $this->course->id, 'Z');
         set_config('defaultuserpreference', 0, 'format_topcoll');
         set_config('defaulttogglepersistence', 1, 'format_topcoll');
 
@@ -489,7 +498,7 @@ final class courseformatrenderer_test extends \advanced_testcase {
         $theoutput .= '<ul class="ctopics ctoggled topics col-sm-12">';
         $theoutput .= self::call_method($this->outputus, 'topcoll_section', [$section1, $this->course, false, null, $toggle]);
         $theoutput .= '</ul></div>';
-        $theoutput .= '<span id="tcdata" class="d-none" data-onetopic="false" data-onetopictoggle="false"'.
+        $theoutput .= '<span id="tcdata" class="d-none" data-onetopic="false" data-onetopictoggle="false"' .
             ' data-defaulttogglepersistence="true"></span>';
 
         $this->assertEquals($theoutput, $thevalue);
@@ -499,7 +508,7 @@ final class courseformatrenderer_test extends \advanced_testcase {
         global $CFG;
 
         $this->init(0);
-        set_user_preference(togglelib::TOPCOLL_TOGGLE.'_' . $this->course->id, null);
+        set_user_preference(togglelib::TOPCOLL_TOGGLE . '_' . $this->course->id, null);
         set_config('defaultuserpreference', 0, 'format_topcoll');
         set_config('defaulttogglepersistence', 1, 'format_topcoll');
         $section0 = $this->courseformat->get_section(0);
@@ -510,7 +519,7 @@ final class courseformatrenderer_test extends \advanced_testcase {
         $theoutput .= '<ul class="ctopics">';
         $theoutput .= self::call_method($this->outputus, 'topcoll_section', [$section0, $this->course, false, 0]);
         $theoutput .= '</ul>';
-        $theoutput .= '<span id="tcdata" class="d-none" data-onetopic="false" data-onetopictoggle="false"'.
+        $theoutput .= '<span id="tcdata" class="d-none" data-onetopic="false" data-onetopictoggle="false"' .
             ' data-defaulttogglepersistence="true"></span>';
 
         $this->assertEquals($theoutput, $thevalue);